FireGuardProject/pyproject.toml
[project]
name = "FireGuardPr"
version = "0.1.0"
description = ""
authors = [
    {name = "h602620",email = "602620@stud.hvl.no"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "dynamic-frcm (>=1.0.1,<2.0.0)",
    "folium (>=0.19.4,<0.20.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "supabase (>=2.13.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
